Fix cucumber failing at first run

+++ features/support/hooks.rb
@@ -1,3 +1,3 @@
 Before do
-  FileUtils.rm_r ('tmp/features')
+  FileUtils.rm_r ('tmp/features') if File.exist?('tmp/features')
 end
